src/test/java/org/cis1200/MarkovChainTest.java
package org.cis1200;

import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;

import java.util.ArrayList;
import java.util.Arrays;

/** Tests for MarkovChain */
public class MarkovChainTest {

    
    @Test
    public void testAddBigram() {
        MarkovChain mc = new MarkovChain();
        mc.addBigram("1", "2");
        assertTrue(mc.chain.containsKey("1"));
        ProbabilityDistribution<String> pd = mc.chain.get("1");
        assertTrue(pd.getRecords().containsKey("2"));
        assertEquals(1, pd.count("2"));
    }

    @Test
    public void testAddBigramWithNullFirst() {
        MarkovChain mc = new MarkovChain();
        try {
            mc.addBigram(null, "second");
        } catch (IllegalArgumentException e) {
            assert true;
        }
    }

    @Test
    public void testAddBigramWithNullSecond() {
        MarkovChain mc = new MarkovChain();
        try {
            mc.addBigram("first", null);
        } catch (IllegalArgumentException e) {
            assert true;
        }
    }

    @Test
    public void testAddBigramWithBothNull() {
        MarkovChain mc = new MarkovChain();
        try {
            mc.addBigram(null, null);
        } catch (IllegalArgumentException e) {
            assert true;
        }
    }




    @Test
    public void testTrain() {
        MarkovChain mc = new MarkovChain();
        String sentence = "1 2 3";
        mc.train(Arrays.stream(sentence.split(" ")).iterator());
        assertEquals(3, mc.chain.size());
        ProbabilityDistribution<String> pd1 = mc.chain.get("1");
        assertTrue(pd1.getRecords().containsKey("2"));
        assertEquals(1, pd1.count("2"));
        ProbabilityDistribution<String> pd2 = mc.chain.get("2");
        assertTrue(pd2.getRecords().containsKey("3"));
        assertEquals(1, pd2.count("3"));
        ProbabilityDistribution<String> pd3 = mc.chain.get("3");
        assertTrue(pd3.getRecords().containsKey(MarkovChain.END_TOKEN));
        assertEquals(1, pd3.count(MarkovChain.END_TOKEN));
    }




   
    @Test
    public void testWalk() {
        
        String[] expectedWords = { "CIS", "1200", "beats", "CIS", "1200", "rocks" };
        MarkovChain mc = new MarkovChain();

        String sentence1 = "CIS 1200 rocks";
        String sentence2 = "CIS 1200 beats CIS 1600";
        mc.train(Arrays.stream(sentence1.split(" ")).iterator());
        mc.train(Arrays.stream(sentence2.split(" ")).iterator());

        mc.reset("CIS"); 
        mc.fixDistribution(new ArrayList<>(Arrays.asList(expectedWords)));

        for (int i = 0; i < expectedWords.length; i++) {
            System.out.println(i);
            assertTrue(mc.hasNext());
            assertEquals(expectedWords[i], mc.next());
        }

    }

}
